fix: fixed bugs

+++ server/server.cpp
@@ -2,7 +2,7 @@
 #include "connection/connection.hpp"
 
 int32_t main() {
-	uint64_t port;
+	int32_t port;
 
 	std::cout << "Write the port: ";
 	std::cin >> port;
